Move profiling/debug functions to separate file to clean up a bit
