+++ GPU/GPUTracking/Global/GPUChainTracking.cxx
@@ -1432,90 +1432,6 @@ int GPUChainTracking::RunChain()
   return 0;
 }
 
-int GPUChainTracking::PrepareProfile()
-{
-#ifdef GPUCA_TRACKLET_CONSTRUCTOR_DO_PROFILE
-  char* tmpMem;
-  GPUFailedMsg(cudaMalloc(&tmpMem, 100000000));
-  processorsShadow()->tpcTrackers[0].mStageAtSync = tmpMem;
-  GPUFailedMsg(cudaMemset(processorsShadow()->tpcTrackers[0].StageAtSync(), 0, 100000000));
-#endif
-  return 0;
-}
-
-int GPUChainTracking::DoProfile()
-{
-#ifdef GPUCA_TRACKLET_CONSTRUCTOR_DO_PROFILE
-  char* stageAtSync = (char*)malloc(100000000);
-  GPUFailedMsg(cudaMemcpy(stageAtSync, processorsShadow()->tpcTrackers[0].StageAtSync(), 100 * 1000 * 1000, cudaMemcpyDeviceToHost));
-  cudaFree(processorsShadow()->tpcTrackers[0].StageAtSync());
-
-  FILE* fp = fopen("profile.txt", "w+");
-  FILE* fp2 = fopen("profile.bmp", "w+b");
-
-  const int bmpheight = 8192;
-  BITMAPFILEHEADER bmpFH;
-  BITMAPINFOHEADER bmpIH;
-  ZeroMemory(&bmpFH, sizeof(bmpFH));
-  ZeroMemory(&bmpIH, sizeof(bmpIH));
-
-  bmpFH.bfType = 19778; //"BM"
-  bmpFH.bfSize = sizeof(bmpFH) + sizeof(bmpIH) + (mConstructorBlockCount * GPUCA_THREAD_COUNT_CONSTRUCTOR / 32 * 33 - 1) * bmpheight;
-  bmpFH.bfOffBits = sizeof(bmpFH) + sizeof(bmpIH);
-
-  bmpIH.biSize = sizeof(bmpIH);
-  bmpIH.biWidth = mConstructorBlockCount * GPUCA_THREAD_COUNT_CONSTRUCTOR / 32 * 33 - 1;
-  bmpIH.biHeight = bmpheight;
-  bmpIH.biPlanes = 1;
-  bmpIH.biBitCount = 32;
-
-  fwrite(&bmpFH, 1, sizeof(bmpFH), fp2);
-  fwrite(&bmpIH, 1, sizeof(bmpIH), fp2);
-
-  int nEmptySync = 0;
-  for (int i = 0; i < bmpheight * mConstructorBlockCount * GPUCA_THREAD_COUNT_CONSTRUCTOR; i += mConstructorBlockCount * GPUCA_THREAD_COUNT_CONSTRUCTOR) {
-    int fEmpty = 1;
-    for (int j = 0; j < mConstructorBlockCount * GPUCA_THREAD_COUNT_CONSTRUCTOR; j++) {
-      fprintf(fp, "%d\t", stageAtSync[i + j]);
-      int color = 0;
-      if (stageAtSync[i + j] == 1) {
-        color = RGB(255, 0, 0);
-      }
-      if (stageAtSync[i + j] == 2) {
-        color = RGB(0, 255, 0);
-      }
-      if (stageAtSync[i + j] == 3) {
-        color = RGB(0, 0, 255);
-      }
-      if (stageAtSync[i + j] == 4) {
-        color = RGB(255, 255, 0);
-      }
-      fwrite(&color, 1, sizeof(int), fp2);
-      if (j > 0 && j % 32 == 0) {
-        color = RGB(255, 255, 255);
-        fwrite(&color, 1, 4, fp2);
-      }
-      if (stageAtSync[i + j]) {
-        fEmpty = 0;
-      }
-    }
-    fprintf(fp, "\n");
-    if (fEmpty) {
-      nEmptySync++;
-    } else {
-      nEmptySync = 0;
-    }
-    (void)nEmptySync;
-    // if (nEmptySync == GPUCA_SCHED_ROW_STEP + 2) break;
-  }
-
-  fclose(fp);
-  fclose(fp2);
-  free(stageAtSync);
-#endif
-  return 0;
-}
-
 int GPUChainTracking::HelperReadEvent(int iSlice, int threadId, GPUReconstructionHelpers::helperParam* par) { return ReadEvent(iSlice, threadId); }
 
 int GPUChainTracking::HelperOutput(int iSlice, int threadId, GPUReconstructionHelpers::helperParam* par)
